feat: LunarTestCase setup + update each test suite
