+++ tests/core/TestCase.php
@@ -3,61 +3,20 @@
 namespace Lunar\Tests\Core;
 
 use Cartalyst\Converter\Laravel\ConverterServiceProvider;
-use Illuminate\Support\Facades\Config;
 use Kalnoy\Nestedset\NestedSetServiceProvider;
-use Lunar\Facades\Taxes;
-use Lunar\LunarServiceProvider;
-use Lunar\Tests\Core\Stubs\TestTaxDriver;
-use Lunar\Tests\Core\Stubs\TestUrlGenerator;
-use Lunar\Tests\Core\Stubs\User;
+use Lunar\Tests\LunarTestCase;
 use Spatie\Activitylog\ActivitylogServiceProvider;
-use Spatie\LaravelBlink\BlinkServiceProvider;
 use Spatie\MediaLibrary\MediaLibraryServiceProvider;
 
-class TestCase extends \Orchestra\Testbench\TestCase
+class TestCase extends LunarTestCase
 {
-    protected function setUp(): void
+    protected function getPackageProviders($app): array
     {
-        parent::setUp();
-        // additional setup
-        Config::set('providers.users.model', User::class);
-        Config::set('lunar.urls.generator', TestUrlGenerator::class);
-        Config::set('lunar.taxes.driver', 'test');
-
-        Taxes::extend('test', function ($app) {
-            return $app->make(TestTaxDriver::class);
-        });
-
-        activity()->disableLogging();
-
-        // Freeze time to avoid timestamp errors
-        $this->freezeTime();
-    }
-
-    protected function getPackageProviders($app)
-    {
-        return [
-            LunarServiceProvider::class,
+        return array_merge(parent::getPackageProviders($app), [
             MediaLibraryServiceProvider::class,
             ActivitylogServiceProvider::class,
             ConverterServiceProvider::class,
             NestedSetServiceProvider::class,
-            BlinkServiceProvider::class,
-        ];
-    }
-
-    protected function getEnvironmentSetUp($app)
-    {
-        // perform environment setup
-    }
-
-    /**
-     * Define database migrations.
-     *
-     * @return void
-     */
-    protected function defineDatabaseMigrations()
-    {
-        $this->loadLaravelMigrations();
+        ]);
     }
 }
